+++ contracts/dca/src/handlers/after_fin_limit_order_retracted.rs
@@ -1,6 +1,8 @@
 use crate::contract::AFTER_FIN_LIMIT_ORDER_WITHDRAWN_FOR_CANCEL_VAULT_REPLY_ID;
 use crate::error::ContractError;
-use crate::state::{create_event, get_vault, update_vault, CACHE, LIMIT_ORDER_CACHE};
+use crate::state::cache::{CACHE, LIMIT_ORDER_CACHE};
+use crate::state::events::create_event;
+use crate::state::vaults::{get_vault, update_vault};
 use crate::vault::Vault;
 use base::events::event::{EventBuilder, EventData};
 use base::helpers::message_helpers::{find_first_attribute_by_key, find_first_event_by_type};
