+++ contracts/dca/src/handlers/cancel_vault.rs
@@ -1,9 +1,9 @@
 use crate::contract::AFTER_FIN_LIMIT_ORDER_RETRACTED_REPLY_ID;
 use crate::error::ContractError;
-use crate::state::{
-    create_event, delete_trigger, get_trigger, get_vault, update_vault, Cache, LimitOrderCache,
-    CACHE, LIMIT_ORDER_CACHE,
-};
+use crate::state::cache::{Cache, LimitOrderCache, CACHE, LIMIT_ORDER_CACHE};
+use crate::state::events::create_event;
+use crate::state::triggers::{delete_trigger, get_trigger};
+use crate::state::vaults::{get_vault, update_vault};
 use crate::validation_helpers::{
     assert_sender_is_admin_or_vault_owner, assert_vault_is_not_cancelled,
 };
