+++ contracts/dca/src/handlers/delete_pair.rs
@@ -1,5 +1,6 @@
+use crate::error::ContractError;
+use crate::state::pairs::PAIRS;
 use crate::validation_helpers::assert_sender_is_admin;
-use crate::{error::ContractError, state::PAIRS};
 use cosmwasm_std::{Addr, DepsMut};
 #[cfg(not(feature = "library"))]
 use cosmwasm_std::{Env, MessageInfo, Response};
