+++ contracts/dca/src/handlers/deposit.rs
@@ -1,5 +1,6 @@
 use crate::error::ContractError;
-use crate::state::{create_event, get_vault, update_vault};
+use crate::state::events::create_event;
+use crate::state::vaults::{get_vault, update_vault};
 use crate::validation_helpers::{
     assert_denom_matches_pair_denom, assert_exactly_one_asset, assert_vault_is_not_cancelled,
 };
