+++ contracts/dca/src/handlers/get_events.rs
@@ -1,6 +1,5 @@
-use crate::{
-    msg::EventsResponse, state::event_store, validation_helpers::assert_page_limit_is_valid,
-};
+use crate::state::events::event_store;
+use crate::{msg::EventsResponse, validation_helpers::assert_page_limit_is_valid};
 use base::events::event::Event;
 use cosmwasm_std::{Deps, StdResult};
 use cw_storage_plus::Bound;
