+++ contracts/dca/src/handlers/get_vault.rs
@@ -1,5 +1,6 @@
-use crate::state::get_vault as fetch_vault;
-use crate::{msg::VaultResponse, state::get_trigger};
+use crate::{
+    msg::VaultResponse, state::triggers::get_trigger, state::vaults::get_vault as fetch_vault,
+};
 use cosmwasm_std::{Addr, Deps, StdError, StdResult, Uint128};
 
 pub fn get_vault(deps: Deps, address: Addr, vault_id: Uint128) -> StdResult<VaultResponse> {
