+++ contracts/dca/src/state/state_helpers.rs
@@ -1,0 +1,8 @@
+use cosmwasm_std::{StdResult, Storage};
+use cw_storage_plus::Item;
+
+pub fn fetch_and_increment_counter(store: &mut dyn Storage, counter: Item<u64>) -> StdResult<u64> {
+    let id = counter.may_load(store)?.unwrap_or_default() + 1;
+    counter.save(store, &id)?;
+    Ok(id)
+}
